fix erros

+++ src/ft_printf.h
@@ -19,7 +19,7 @@ typedef struct format_parser
 
 } format_parser;
 
-#define INIT_PARSER(P) format_parser P = {.errors_count = 0, .flag = ' ', .precision = -1, .specifier = ' ', .width = -1, .is_dynamic_precision = 0, .is_dynamic_wdith = 0};
+#define INIT_PARSER(P) format_parser P = { .errors_count = 0, .flag = ' ', .precision = -1, .specifier = ' ', .width = -1, .is_dynamic_precision = 0, .is_dynamic_wdith = 0};
 
 int ft_printf(char *args, ...);
 void parse_flags(int *i, char *str, format_parser *parser);
@@ -37,13 +37,13 @@ void output_x_specifier(va_list *parms_arry, format_parser *parser);
 void output_X_specifier(va_list *parms_arry, format_parser *parser);
 
 void write_nb(int nb, int precision, char flag);
-void write_nb_u(unsigned long nb, int precision);
+void write_nb_u(long long nb, int precision);
 void write_str(char *str, int precision);
 void write_str_x(char *str, int precision);
 void write_flags(char c, char *str,int width, int precision);
 void write_flag_p(int size);
 void output_int_flag(int nb, int width, int precision, char flag);
-void output_pointer_flag(unsigned int ptr, int width, int lenght, char flag);
+void output_pointer_flag(long long int ptr, int width, int lenght, char flag);
 void output_s_flags(format_parser *parser, char *str, int precision, int width);
 void output_x_flags(format_parser *parser, char *str, int precision, int width);
 int is_valid_specifier(char c);
@@ -51,7 +51,7 @@ int is_valid_specifier(char c);
 int ft_strlen(const char *str);
 void ft_putchar(char c);
 void ft_putstr(char *s);
-void ft_putnbr(long nb);
+void ft_putnbr(long long nb);
 int ft_isdigit(int c);
 int check_base(char *base);
 char *ft_itoa(long long nbr, char *base);
